Moved the new files to the events package.
